UpdateRouts

+++ routes/web.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\Admin\BookController;
 use App\Http\Controllers\Visitor\VisitorBookController;
+use App\Http\Controllers\Admin\UserController;
 
 
 /*
@@ -30,26 +31,20 @@ require __DIR__.'/auth.php';
 //     return view('visitor.index');
 // });
 
-Route::prefix('admin')->group(function () {
-    Route::get('books', [BookController::class, 'index'])->name('admin.books.index');
-    Route::get('books/create', [BookController::class, 'create'])->name('admin.books.create');
-    Route::post('books', [BookController::class, 'store'])->name('admin.books.store');
-    Route::get('books/{book}', [BookController::class, 'show'])->name('admin.books.show');
-    Route::get('books/{book}/edit', [BookController::class, 'edit'])->name('admin.books.edit');
-    Route::patch('books/{book}', [BookController::class, 'update'])->name('admin.books.update');
-    Route::delete('books/{book}', [BookController::class, 'destroy'])->name('admin.books.destroy');
-    Route::get('books/{book}/delete', [BookController::class, 'delete'])->name('admin.books.delete');
-    Route::post('books/sell', [BookController::class, 'sellBook'])->name('admin.books.sell');
-});
 
 
-Route::get('/books', [BookController::class, 'index'])->name('books.index');
-Route::get('/books/create', [BookController::class, 'create'])->name('books.create');
-Route::post('/books', [BookController::class, 'store'])->name('books.store');
-Route::get('/books/{book}', [BookController::class, 'show'])->name('books.show');
-Route::get('/books/{book}/edit', [BookController::class, 'edit'])->name('books.edit');
-Route::put('/books/{book}', [BookController::class, 'update'])->name('books.update');
-Route::delete('/books/{book}', [BookController::class, 'destroy'])->name('books.destroy');
+
+Route::get('admin/books', [BookController::class, 'index'])->name('admin.books.index');
+Route::get('admin/books/create', [BookController::class, 'create'])->name('admin.books.create');
+Route::post('admin/books', [BookController::class, 'store'])->name('admin.books.store');
+Route::get('admin/books/{book}', [BookController::class, 'show'])->name('admin.books.show');
+Route::get('admin/books/{book}/edit', [BookController::class, 'edit'])->name('admin.books.edit');
+Route::put('admin/books/{book}', [BookController::class, 'update'])->name('admin.books.update');
+Route::delete('admin/books/{book}', [BookController::class, 'destroy'])->name('admin.books.destroy');
+Route::get('admin/user-purchases', [UserController::class, 'showUserPurchases'])->name('admin.user.purchases');
+
+
+
 
 //visitor Routs
 Route::get('/', [VisitorBookController::class, 'index'])->name('home');
